validator-api: extract out some calculations
